refactor: move tests from src/__tests__/ to top-level test/ folder
- Relocate all test files to test/ directory per project convention
- Update all import paths to reference src/ modules correctly
